services/reco-python/requirements.txt
fastapi==0.115.6
uvicorn==0.30.6
psycopg2-binary==2.9.9
